replace implicit conversion with Scala 3 extension
